+++ CHANGELOG.md
@@ -48,6 +48,7 @@ v2.0
 - Fix FLEXIAPI-325 Add endpoints to send the password reset and provisioning emails
 - Fix FLEXIAPI-332 Check if the first line was untouched and that the number of columns is exact on each lines
 - Fix FLEXIAPI-329 Use correct routes for accounts devices
+- Fix FLEXIAPI-330 Remove the ConfirmedRegistration email and related code
 
 v1.6
 ----